backend/exparenza_backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
sqlparse==0.5.3
whitenoise==6.11.0
